requirements.txt
langchain
streamlit
python-dotenv
langchain-huggingface`